refactor: 💡 swc, removed babel
updated to be more next12

style: 💄 added sass support

fix: 🐛 missed a file rename
