+++ jest.config.js
@@ -1,4 +1,17 @@
-module.exports = {
+const nextJest = require("next/jest");
+const createJestConfig = nextJest({
+  // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
+  dir: "./",
+});
+
+// Add any custom config to be passed to Jest
+const customJestConfig = {
+  // Add more setup options before each test is run
+  // setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
+  // if using TypeScript with a baseUrl set to the root directory then you need the below for alias' to work
+  moduleDirectories: ["node_modules", "<rootDir>/"],
+  testEnvironment: "jest-environment-jsdom",
+  rootDir: "./",
   // Automatically clear mock calls and instances between every test
   clearMocks: true,
 
@@ -17,7 +30,8 @@ module.exports = {
       "<rootDir>/tests/fileMock.js",
     "\\.(css|less)$": "identity-obj-proxy",
   },
-  testEnvironment: "jsdom",
   // The glob patterns Jest uses to detect test files
   testMatch: ["**/tests/**/*.test.[jt]s?(x)"],
 };
+
+module.exports = createJestConfig(customJestConfig);
