+++ package.json
@@ -19,16 +19,16 @@
   },
   "dependencies": {
     "axios": "^0.21.1",
-    "next": "11.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-query": "^3.18.1"
+    "next": "^12.2.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-query": "^3.18.1",
+    "sass": "^1.54.0"
   },
   "devDependencies": {
     "@testing-library/cypress": "^7.0.6",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
-    "babel-jest": "^27.0.6",
     "cypress": "^7.7.0",
     "eslint": "^7.31.0",
     "eslint-config-next": "^11.0.1",
@@ -45,9 +45,11 @@
     "eslint-plugin-security": "^1.4.0",
     "husky": "^7.0.1",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.0.6",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
     "lint-staged": "^11.1.1",
     "msw": "^0.32.0",
+    "next-images": "^1.8.4",
     "npm-run-all": "^4.1.5",
     "pre-commit": "^1.2.2",
     "prettier": "^2.3.2",
