+++ pages/solar-flares.js
@@ -2,7 +2,7 @@
 
 import Head from "next/head";
 import Image from "next/image";
-import styles from "../styles/Home.module.css";
+import styles from "../styles/Home.module.scss";
 import { useEffect, useState } from "react";
 import axios from "axios";
