+++ styles/globals.scss
@@ -1,3 +1,5 @@
+$primaryColor: tomato;
+
 html,
 body {
   padding: 0;
@@ -7,7 +9,7 @@ body {
 }
 
 a {
-  color: inherit;
+  color: $primaryColor;
   text-decoration: none;
 }
